+++ sdk/keyvault/Azure.Security.KeyVault.Certificates/src/CertificateCreateParameters.cs
@@ -12,20 +12,20 @@ namespace Azure.Security.KeyVault.Certificates
         private const string AttributesPropertyName = "attributes";
         private const string EnabledPropertyName = "enabled";
         private const string TagsPropertyName = "tags";
-        private const string PreserveCertOrderPropertyName = "preserveCertOrder";
+        private const string PreserveCertificateOrderPropertyName = "preserveCertOrder";
 
         private static readonly JsonEncodedText s_policyPropertyNameBytes = JsonEncodedText.Encode(PolicyPropertyName);
         private static readonly JsonEncodedText s_attributesPropertyNameBytes = JsonEncodedText.Encode(AttributesPropertyName);
         private static readonly JsonEncodedText s_enabledPropertyNameBytes = JsonEncodedText.Encode(EnabledPropertyName);
         private static readonly JsonEncodedText s_tagsPropertyNameBytes = JsonEncodedText.Encode(TagsPropertyName);
-        private static readonly JsonEncodedText s_preserveCertOrderPropertyNameBytes = JsonEncodedText.Encode(PreserveCertOrderPropertyName);
+        private static readonly JsonEncodedText s_preserveCertificateOrderPropertyNameBytes = JsonEncodedText.Encode(PreserveCertificateOrderPropertyName);
 
-        public CertificateCreateParameters(CertificatePolicy policy, bool? enabled, IDictionary<string, string> tags, bool? preserveCertOrder = null)
+        public CertificateCreateParameters(CertificatePolicy policy, bool? enabled, IDictionary<string, string> tags, bool? preserveCertificateOrder = null)
         {
             Policy = policy;
             Enabled = enabled;
             Tags = tags;
-            PreserveCertOrder = preserveCertOrder;
+            PreserveCertificateOrder = preserveCertificateOrder;
         }
 
         public CertificatePolicy Policy { get; }
@@ -34,7 +34,7 @@ namespace Azure.Security.KeyVault.Certificates
 
         public IDictionary<string, string> Tags { get; }
 
-        public bool? PreserveCertOrder { get; }
+        public bool? PreserveCertificateOrder { get; }
 
         void IJsonSerializable.WriteProperties(Utf8JsonWriter json)
         {
@@ -68,9 +68,9 @@ namespace Azure.Security.KeyVault.Certificates
                 json.WriteEndObject();
             }
 
-            if (PreserveCertOrder.HasValue)
+            if (PreserveCertificateOrder.HasValue)
             {
-                json.WriteBoolean(s_preserveCertOrderPropertyNameBytes, PreserveCertOrder.Value);
+                json.WriteBoolean(s_preserveCertificateOrderPropertyNameBytes, PreserveCertificateOrder.Value);
             }
         }
     }