+++ sdk/keyvault/Azure.Security.KeyVault.Certificates/src/ImportCertificateOptions.cs
@@ -21,7 +21,7 @@ namespace Azure.Security.KeyVault.Certificates
         private static readonly JsonEncodedText s_attributesPropertyNameBytes = JsonEncodedText.Encode("attributes");
         private static readonly JsonEncodedText s_enabledPropertyNameBytes = JsonEncodedText.Encode("enabled");
         private static readonly JsonEncodedText s_tagsPropertyNameBytes = JsonEncodedText.Encode("tags");
-        private static readonly JsonEncodedText s_preserveCertOrderPropertyNameBytes = JsonEncodedText.Encode("preserveCertOrder");
+        private static readonly JsonEncodedText s_preserveCertificateOrderPropertyNameBytes = JsonEncodedText.Encode("preserveCertOrder");
 
         private Dictionary<string, string> _tags;
 
@@ -84,7 +84,7 @@ namespace Azure.Security.KeyVault.Certificates
         /// Gets or sets a value indicating whether the certificate chain preserves its original order.
         /// The default value is false, which sets the leaf certificate at index 0.
         /// </summary>
-        public bool? PreserveCertOrder { get; set; }
+        public bool? PreserveCertificateOrder { get; set; }
 
         void IJsonSerializable.WriteProperties(Utf8JsonWriter json)
         {
@@ -136,9 +136,9 @@ namespace Azure.Security.KeyVault.Certificates
                 json.WriteEndObject();
             }
 
-            if (PreserveCertOrder.HasValue)
+            if (PreserveCertificateOrder.HasValue)
             {
-                json.WriteBoolean(s_preserveCertOrderPropertyNameBytes, PreserveCertOrder.Value);
+                json.WriteBoolean(s_preserveCertificateOrderPropertyNameBytes, PreserveCertificateOrder.Value);
             }
         }
     }